word_master/requirements.txt
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.3
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
pandas==2.0.0
psycopg2==2.9.6
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
tzdata==2023.3
Werkzeug==2.2.3
